// apps/squoolr-v2/api/src/utils/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  PARENT = 'PARENT',
  STUDENT = 'STUDENT',
  TEACHER = 'TEACHER',
  REGISTRY = 'REGISTRY',
  COORDINATOR = 'COORDINATOR',
  CONFIGURATOR = 'CONFIGURATOR',
}

export enum MetadataEnum {
  ROLES = 'roles',
  IS_PUBLIC = 'isPublic',
  IS_PRIVATE = 'isPrivate',
}

export enum StaffRole {
  TEACHER = 'TEACHER',
  REGISTRY = 'REGISTRY',
  COORDINATOR = 'COORDINATOR',
  CONFIGURATOR = 'CONFIGURATOR',
}

export enum KeyRole {
  STAFF = 'STAFF',
  ADMIN = 'ADMIN',
  PARENT = 'PARENT',
  STUDENT = 'STUDENT',
}